// packages/db-client/src/instantdb-types.ts
export type DBIdent =
  | [string, string, string]
  | [string, string, string, boolean];

export type CheckedDataType = "string" | "number" | "boolean" | "date";

export interface DBAttr {
  id: string;
  "forward-identity": DBIdent;
  "reverse-identity"?: DBIdent;
  "index?": boolean;
  "unique?": boolean;
  "required?"?: boolean;
  "primary?"?: boolean | undefined;
  cardinality: "one" | "many";
  "value-type": "ref" | "blob";
  "inferred-types"?: Array<"string" | "number" | "boolean" | "json">;
  catalog?: "user" | "system";
  "checked-data-type"?: CheckedDataType;
  "on-delete"?: "cascade";
  "on-delete-reverse"?: "cascade";
}

export interface SchemaNamespace {
  id: string;
  name: string;
  attrs: SchemaAttr[];
}

export interface SchemaNamespaceMap {
  id: string;
  name: string;
  attrs: Record<string, SchemaAttr>;
}

export interface SchemaAttr {
  id: string;
  isForward: boolean;
  namespace: string;
  name: string;
  type: "ref" | "blob";
  isIndex: boolean;
  isUniq: boolean;
  isRequired?: boolean;
  isPrimary?: boolean | undefined;
  cardinality: "one" | "many";
  linkConfig: {
    forward: {
      id: string;
      namespace: string;
      attr: string;
      nsMap: SchemaNamespaceMap | undefined;
    };
    reverse:
      | {
          id: string;
          namespace: string;
          attr: string;
          nsMap: SchemaNamespaceMap | undefined;
        }
      | undefined;
  };
  inferredTypes?: Array<"string" | "number" | "boolean" | "json">;
  catalog?: "user" | "system";
  checkedDataType?: CheckedDataType;
  sortable: boolean;
  onDelete?: "cascade";
  onDeleteReverse?: "cascade";
}
// ...
export type SearchFilterOp =
  | "="
  | "$ilike"
  | "$like"
  | "$gt"
  | "$lt"
  | "$isNull";
export type SearchFilter = [string, SearchFilterOp, unknown];